unify stac utils
